requirements.txt
flake8
pycodestyle
mypy
ruff
pylint
requests==2.32.3
PyYAML==6.0.2

